requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.8.2
pydantic-settings==2.4.0
chromadb==0.5.5
sentence-transformers==3.0.1
pypdf==4.3.1
numpy==1.26.4
ulid-py==1.1.0
httpx==0.27.0
openai==1.42.0
google-generativeai==0.7.2
onnxruntime
PyMuPDF
docx